m68k: speed up task context switches
